ResearchAgent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Research Agent using Pydantic-AI with Logfire instrumentation"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=0.0.14",
    "logfire>=0.50.0",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
